changed last layer from Softmaxwithloss to Softmax

+++ yonatan/how_many_categorized.py
@@ -74,7 +74,7 @@ def how_many(argv):
         deleted = sum_of_all_already_seen - sum_of_all
 
         for key, value in dress_length_dict.iteritems():
-            print '{0}: {1}, percent: {2}%'.format(key, value, round(float(value) / sum_of_all, 2) * 100)
+            print '{0}: {1}, percent: {2}%'.format(key, value, int(round(float(value) / sum_of_all, 2) * 100))
 
     # men shirt sleeve #
     elif args.input_file == 'men_shirt_sleeve':
